add test for kms_key and rename file to match
