+++ src/storage/examples/src/objects/upload_with_kms_key.rs
